requirements.txt
kafka-python==2.0.2
watchdog==2.1.8
